requirements.txt
streamlit~=1.26.0
Pillow~=9.5.0
clarifai-grpc
python-dotenv~=1.0.0
numpy~=1.25.2